+++ back/src/gmail/gmail.controller.ts
@@ -3,6 +3,7 @@ import {
   Controller,
   Delete,
   Get,
+  NotFoundException,
   Post,
   Query,
   Req,
@@ -136,16 +137,20 @@ export class GmailController {
     description: 'The subscription has been successfully deleted.',
   })
   @UseGuards(AuthGuard('jwt'))
-  async deleteSubscription(@Req() req, @Res() res, @Query('id') id: string) {
-    try {
-      await this.gmailService.deleteSubscription(
-        id,
-        await this.authService.getValidGmailToken(req.user.id)
-      );
-      return res.status(200).send({ message: 'Subscription deleted' });
-    } catch (error) {
-      console.error('Error deleting subscription:', error);
-      return res.status(500).send({ message: 'Failed to delete subscription' });
+  async deleteSubscription(@Req() req, @Query('id') id: number) {
+    const hook = await this.hooksRepository.findOne({
+      where: { id: id, userId: req.user.id, service: 'gmail' },
+    });
+
+    if (!hook) {
+      throw new NotFoundException('Hook not found');
     }
+
+    await this.gmailService.deleteSubscription(
+      hook.webhookId,
+      await this.authService.getValidGmailToken(req.user.id)
+    );
+
+    return { message: 'Subscription deleted' };
   }
 }
